+++ NEWS.md
@@ -8,6 +8,7 @@
 - Simplification: No need to define `format`, `echo`, and `warning` in YAML header of `survey.qmd` anymore. The survey can have a clean start and runs fine without the YAML header.
 - Simplification: No need to define `sd_nav()` for each page. It's auto-injected unless `sd_nav()` or `sd_close()` is explicitly defined in the page with or without parameters.
 - New `settings.yml` content structure with 3 sections: `theme-settings`, `survey-settings`, and `system-messages`. They appear with our without YAML header of `survey.qmd` and are verbose (meaning all keys and values always appear).
+- Progress bar update: Now since previous button is supported, the progress bar will update accordingly when user goes back and forth among pages.
 
 # surveydown 0.14.0